+++ lib/internal/modules/esm/loader.js
@@ -31,6 +31,7 @@ const {
   ERR_UNKNOWN_MODULE_FORMAT
 } = require('internal/errors').codes;
 const { pathToFileURL, isURLInstance, URL } = require('internal/url');
+const { emitExperimentalWarning } = require('internal/util');
 const {
   isAnyArrayBuffer,
   isArrayBufferView,
@@ -52,6 +53,13 @@ const { getOptionValue } = require('internal/options');
 const {
   fetchModule,
 } = require('internal/modules/esm/fetch_module');
+
+
+/**
+ * Prevent the specifier resolution warning from being printed twice
+ */
+let emittedSpecifierResolutionWarning = false;
+
 
 /**
  * An ESMLoader instance is used as the main entry point for loading ES modules.
@@ -106,6 +114,22 @@ class ESMLoader {
    * Methods which translate input code or other information into ES modules
    */
   translators = translators;
+
+  constructor() {
+    if (getOptionValue('--experimental-loader')) {
+      emitExperimentalWarning('Custom ESM Loaders');
+    }
+    if (getOptionValue('--experimental-network-imports')) {
+      emitExperimentalWarning('Network Imports');
+    }
+    if (getOptionValue('--experimental-specifier-resolution') === 'node' && !emittedSpecifierResolutionWarning) {
+      process.emitWarning(
+        'The Node.js specifier resolution flag is experimental. It could change or be removed at any time.',
+        'ExperimentalWarning'
+      );
+      emittedSpecifierResolutionWarning = true;
+    }
+  }
 
   static pluckHooks({
     globalPreload,
